requirements.txt
requests>=2.31.0
urllib3>=2.2.0
Flask>=3.0.0
pydantic>=2.8.0
python-dotenv>=1.0.1
# Optional: vector index
faiss-cpu>=1.8.0.post1; platform_machine != "aarch64"
# Optional: Google Generative AI for embeddings/tools
google-generativeai>=0.7.2
